Dynamic clamp argument parsing for lights/groups.
Usually clamp works on pre-defined subcommands. In our case
we would like to use dynamic subcommands, so that we can run:

  # lolo light1 on
  # lolo group1 off

Clamp compares the subcommand passed on the command line with
the string defined in the SubCommand code definition. By hijacking
the string comparison and replying true/false at will, we can
dynamically chose if the subcommand matches the argument or not.

This way we can iterate over all lights and groups and look for
one that matches the command line argument. If it matches return
yes and set a global variable defining that light or group.

Unfortunately I have not found a way to do that without the
global variable.